add relations
